+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- `mpvd push -p`/`--playlist` flag to push tracks to an XSPF playlist file
+  instead of the current playlist.
+
 ## [2.5.0] - 2026-09-20
 
 ### Added